Moved OSInfoBackend to its own file. Cleanups
